like

+++ database/migrations/2023_04_10_112611_bill.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('bills', function (Blueprint $table) {
             $table->id();
             $table->foreignIdFor(Customer::class)->constrained()->cascadeOnDelete();
-            $table->bigInteger('active');
+            $table->integer('active');
             $table->timestamps();
         });
     }
